Add option to disable cache auto sweep, fix passing non array cache option

+++ src/Discord/Helpers/CacheWrapper.php
@@ -63,7 +63,7 @@ class CacheWrapper
     protected $prefix;
 
     /**
-     * @var callable Callback flusher
+     * @var ?callable Callback flusher
      */
     protected $flusher;
 
@@ -90,32 +90,36 @@ class CacheWrapper
 
         $this->prefix = implode($separator, [substr(strrchr($this->class, '\\'), 1)] + $vars).$separator;
 
-        // Flush every heartbeat ack
-        $this->flusher = function ($time, Discord $discord) {
-            $flushing = 0;
-            foreach ($this->items as $key => $item) {
-                if ($item === null) {
-                    // Item was removed from memory, delete from cache
-                    $this->delete($key);
-                    $flushing++;
-                } elseif ($item instanceof Part) {
-                    // Skip ID related to Bot
-                    if ($key != $discord->id) {
-                        // Item is no longer used other than in the repository, weaken so it can be garbage collected
-                        $this->items[$key] = WeakReference::create($item);
+        if (! empty($discord->options['cacheSweep'])) {
+            // Flush every heartbeat ack
+            $this->flusher = function ($time, Discord $discord) {
+                $flushing = 0;
+                foreach ($this->items as $key => $item) {
+                    if ($item === null) {
+                        // Item was removed from memory, delete from cache
+                        $this->delete($key);
+                        $flushing++;
+                    } elseif ($item instanceof Part) {
+                        // Skip ID related to Bot
+                        if ($key != $discord->id) {
+                            // Item is no longer used other than in the repository, weaken so it can be garbage collected
+                            $this->items[$key] = WeakReference::create($item);
+                        }
                     }
                 }
-            }
-            if ($flushing) {
-                $this->discord->getLogger()->debug('Flushing repository cache', ['count' => $flushing, 'class' => $this->class]);
-            }
-        };
-        $discord->on('heartbeat-ack', $this->flusher);
+                if ($flushing) {
+                    $this->discord->getLogger()->debug('Flushing repository cache', ['count' => $flushing, 'class' => $this->class]);
+                }
+            };
+            $discord->on('heartbeat-ack', $this->flusher);
+        }
     }
 
     public function __destruct()
     {
-        $this->discord->removeListener('heartbeat-ack', $this->flusher);
+        if ($this->flusher) {
+            $this->discord->removeListener('heartbeat-ack', $this->flusher);
+        }
     }
 
     /**
